fix: Resolves two errors thrown by `update_po_list()` (#87)
* fix: Resolves datetime error with `update_po_list()`
Fixes the datetime error thrown by pandas when trying to call 
`ContractManagement.update_contract_list()` and 
`ContractManagement.update_po_list()` if one of the dates was NULL

* fix: Fixes json encoding bug in update_po_list()
A json serialization bug in `ContractManagement.update_po_list()` was 
preventing the insertion of new list items. The cause of this bug was 
that some release numbers were returned from CitiBuy as decimal values 
instead of integers and couldn't be serialized for the payload that was 
passed to the batch upsert post method. This commit fixes that issue by 
casting the Release Number column as an "int64" type.

Fixes: #86