fix: unnecessary updates in update list methods (#91)
* fix: bug in update_vendor_list() that produced unnecessary updates
ContractManagement.update_vendor_list() was producing unnecessary 
updates because blank values in CitiBuy were set as empty strings ('') 
while blank values in SharePoint were read as NaN values. This fix uses 
fillna() to set all NA values as empty strings to avoid creating false 
positives in the update list

* fix: Filters out Open Market POs from update_contract_list()
A bug previously included Open Market POs in the list of contracts that 
would be inserted or updated in 
ContractManagement.update_contract_list() This commit fixes this bug by 
excluding Open Market POs from the contract list that is extracted in 
ContractManagement.get_citibuy_data()

* fix: excludes vendor col from contracts to prevent false updates
Because the vendor column in SharePoint data included the name of the 
Vendor, but the vendor column in CitiBuy data was the vendor id, 
ContractManagement.update_contract_list() was flagging any contract with 
a vendor as an item that needed to be updated. This bug fix filters 
Vendor column out of contract data before passing it to 
_detect_changes() to avoid those false positives

* fix: avoids unnecessary updates in update_po_list()
Previously update_po_list() was produce a lot of false positive updates 
because the values of the vendor and contact lookup fields from CitiBuy 
didn't match those from SharePoint despite looking up to the correct 
vendor/contract. This commit fixes those issues to avoid performing 
unnecessary updates

* fix: prevents accidental inserts when updating current contracts and POs
Previously if there were multiple contracts with the same PO Number in 
CitiBuy, these would be added to the inserts attributes of the 
BatchedChanges object returned by self._detect_changes() even though 
only existing records should be updated in that step. This commit resets 
changes.inserts to an empty list [] to prevent accidental inserts.

* feat: Adds Agency to the list of contract fields updated by the script

* fix: sorts citibuy data by po_nbr, release_nbr, and agency
This ensures that if a PO has two master blanket contracts associated 
with it, we preserve the one that is for DGS.

closes: #90 
closes: #88