fix: Wrong filter condition in `Citibuy.get_invoices()` (#105)
* feat: Adds modified to Invoice table
Adds the modified field to the Invoice table so it can be used in 
`CitiBuy.get_invoices()` to filter for invoices that were modified in the 
last 45 days. The previous strategy of using the `InvoiceStatusHistory`
excluded invoices that were recently paid.

* feat: Returns "modified" field in invoice data exported from CitiBuy

* refactor: Removes commented out portions of CitiBuy.get_invoices()
This removes the sections of `CitiBuy.get_invoices()` that previously 
joined the Invoice table to the InvoiceStatusHistory table to filter by 
the last status update. These lines were commented out because recently 
paid invoices don't have an entry in the InvoiceStatusHistory table for 
when their status was updated to paid.

closes: #104